pyproject.toml
[project]
name = "agent-dev"
version = "0.1.0"
description = "Minimal agent loop with lifecycle hooks"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.28.0",
    "loguru>=0.7.0",
    "pydantic>=2.0",
    "python-dotenv>=1.2.2",
    "wechat-clawbot>=0.3.0",
    "hermes-agent",
]

[tool.uv.sources]
hermes-agent = { git = "https://github.com/NousResearch/hermes-agent.git" }
